Dannys branch

+++ Backend/WebAPI/Controllers/AssessmentsController.cs
@@ -1,3 +1,4 @@
+using System.IO;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using Models.Diagnosis;
@@ -23,7 +24,7 @@ namespace WebAPI.Controllers
             {
                 List<Assessment> assessment = _repo.GetAll().ToList();
                 if (assessment.Count == 0)
-                    throw new Exception("No data found");
+                    throw new FileNotFoundException("No data found");
                 return Ok(_repo.GetAll());
             }
             catch (Exception e)
@@ -40,7 +41,7 @@ namespace WebAPI.Controllers
             try
             {
                 if (_repo.GetById(id) == null)
-                    throw new Exception("Invalid Id");
+                    throw new ArgumentNullException("Invalid Id");
                 return Ok(_repo.GetById(id));
             }
             catch (Exception e)
@@ -56,7 +57,7 @@ namespace WebAPI.Controllers
         {
             try
             {   if (_repo.SearchByDiagnosisId(id) == null)
-                    throw new Exception("Invaild Id");
+                    throw new ArgumentNullException("Invaild Id");
                 return Ok(_repo.SearchByDiagnosisId(id));
             }
             catch (Exception e)
@@ -73,7 +74,7 @@ namespace WebAPI.Controllers
             try
             {
                 if (p_Assessment == null)
-                    throw new Exception("Invalid data!");
+                    throw new ArgumentNullException("Invalid data!");
                 _repo.Create(p_Assessment);
                 _repo.Save();
                 return Created("Assessment/Add", p_Assessment);
@@ -91,8 +92,9 @@ namespace WebAPI.Controllers
         {
             try
             {
-               
                 var topic = _repo.GetById(id);
+                if (topic == null)
+                    throw new ArgumentNullException("Delete failed!");
                 _repo.Delete(topic);
                 _repo.Save();
                 return Ok();
@@ -110,7 +112,6 @@ namespace WebAPI.Controllers
         {
             try
             {
-                 
                 p_Assessment.Id = id;
                 _repo.Update(p_Assessment);
                 _repo.Save();
@@ -122,8 +123,5 @@ namespace WebAPI.Controllers
                 return BadRequest("Invalid Assessments Update");
             }
         }
-
-
-        
     }
 }
